fn comments (exec)
