+++ src/execute/execute.c
@@ -1,5 +1,11 @@
 #include "minishell.h"
 
+/*
+ * Updates cmd->fdin and cmd->fdout if the current t_cmd contains an infile
+ * or an outfile.
+ * 
+ * Exit on: pipe failure, open failure
+ */
 static void	setup_io(t_cmd *cmd, t_cmd **cmd_lst)
 {
 	if (cmd->next)
@@ -15,7 +21,7 @@ static void	setup_io(t_cmd *cmd, t_cmd **cmd_lst)
 		if (cmd->fdin == -1)
 			exit_exec(cmd_lst, cmd->infile);
 	}
-	if (cmd->next) // if we created a pipe in this loop
+	if (cmd->next)
 		cmd->fdout = cmd->pipe[1];
 	else if (cmd->outfile)
 	{
@@ -25,18 +31,32 @@ static void	setup_io(t_cmd *cmd, t_cmd **cmd_lst)
 	}
 }
 
+/*
+ * Closes the uncessary file descriptors for the current t_cmd:
+ * - cmd->fdin if there is an infile
+ * - cmd->outfile if there is an outfile
+ * - pipe's write end if we created a pipe in this cmd
+ * - pipe's read end if we created a pipe in the previous cmd
+ * 
+ * Exit on: none
+ */
 static void	cleanup_io(t_cmd *cmd)
 {
 	if (cmd->infile)
 		close(cmd->fdin);
 	if (cmd->outfile)
 		close(cmd->fdout);
-	if (cmd->next) // if we created a pipe in this loop
-		close(cmd->pipe[1]); // we won't write in pipe in next loop
-	if (cmd->prev) // if we created a pipe in the previous loop
-		close(cmd->prev->pipe[0]); // we finished reading in previous pipe
+	if (cmd->next)
+		close(cmd->pipe[1]);
+	if (cmd->prev)
+		close(cmd->prev->pipe[0]);
 }
 
+/*
+ * Execute each t_cmd of the list one by one.
+ *
+ * Exit on: incorrect input, function call exit
+ */
 void	execute_cmd_lst(t_cmd **cmd_lst, char **envp)
 {
 	t_cmd		*cmd;
@@ -54,7 +74,7 @@ void	execute_cmd_lst(t_cmd **cmd_lst, char **envp)
 		if (builtin && builtin->affects_state)
 			run_builtin(builtin, cmd->args, cmd_lst);
 		else
-			run_executable(builtin, cmd, envp, cmd_lst);
+			run_in_child_process(builtin, cmd, envp, cmd_lst);
 		cleanup_io(cmd);
 		cmd = cmd->next;
 	}
